Add portable AOT support on x86
TR::Compiler->target contains the host environment while TR::Compiler->relocatableTarget contains
the target environment that AOT compilations will use. TR::Compiler->relocatableTarget is set according
to the presence of SCC, portable options and Docker containers. Then at the beginning of every compilation
(in the constructor) we set the target to be TR::Compiler->target if JIT compilation and
TR::Compiler->relocatableTarget if AOT compilation.
The following scenarios were considered:

- When there is no existing SCC and no portable option specified we use the host processor for AOT
- When there is no existing SCC and has portable option specified, we use the hand-picked portable processor for AOT
- When there is no existing SCC and in container, we use the hand-picked portable processor for AOT
- When there is existing SCC and it passes processor compatibility check, we use the processor stored in the SCC
(ignoring portable options and whether we are in container or not)
- JITServer is not affected. Meaning JITServer will still use JITClient's processor information for all of its
compilations regardless of the portable/inContainer/SCC options

depends on: eclipse/omr#4861
Issue: #7966

Signed-off-by: Harry Yu <harryyu1994@gmail.com>